+++ bugsnag-android-core/src/main/java/com/bugsnag/android/BugsnagEventMapper.kt
@@ -17,7 +17,7 @@ internal class BugsnagEventMapper(
 
     @Suppress("UNCHECKED_CAST")
     internal fun convertToEventImpl(map: Map<in String, Any?>, apiKey: String): EventInternal {
-        val event = EventInternal(apiKey)
+        val event = EventInternal(apiKey, logger)
 
         // populate exceptions. check this early to avoid unnecessary serialization if
         // no stacktrace was gathered.