+++ bugsnag-android-core/src/main/java/com/bugsnag/android/internal/InternalMetrics.kt
@@ -16,4 +16,8 @@ interface InternalMetrics {
     fun notifyAddCallback(callback: String)
 
     fun notifyRemoveCallback(callback: String)
+
+    fun setMetadataTrimMetrics(stringsTrimmed: Int, charsRemoved: Int)
+
+    fun setBreadcrumbTrimMetrics(breadcrumbsRemoved: Int, bytesRemoved: Int)
 }