+++ bugsnag-android-core/src/main/java/com/bugsnag/android/internal/InternalMetricsImpl.kt
@@ -5,18 +5,31 @@ import com.bugsnag.android.NdkPluginCaller
 class InternalMetricsImpl : InternalMetrics {
     private val configDifferences = hashMapOf<String, Any>()
     private val callbackCounts = hashMapOf<String, Int>()
+    private var metadataStringsTrimmedCount = 0
+    private var metadataCharsTruncatedCount = 0
+    private var breadcrumbsRemovedCount = 0
+    private var breadcrumbBytesRemovedCount = 0
 
     override fun toJsonableMap(): Map<String, Any> {
-        return mapOf(
-            "config" to configDifferences,
-            "callbacks" to allCallbacks()
-        )
+        val system = listOfNotNull(
+            if (metadataStringsTrimmedCount > 0) "stringsTruncated" to metadataStringsTrimmedCount else null,
+            if (metadataCharsTruncatedCount > 0) "stringCharsTruncated" to metadataCharsTruncatedCount else null,
+            if (breadcrumbsRemovedCount > 0) "breadcrumbsRemoved" to breadcrumbsRemovedCount else null,
+            if (breadcrumbBytesRemovedCount > 0) "breadcrumbBytesRemoved" to breadcrumbBytesRemovedCount else null,
+        ).toMap()
+        val callbacks = allCallbacks()
+
+        return listOfNotNull(
+            if (configDifferences.isNotEmpty()) "config" to configDifferences else null,
+            if (callbacks.isNotEmpty()) "callbacks" to callbacks else null,
+            if (system.isNotEmpty()) "system" to system else null,
+        ).toMap()
     }
 
     override fun setConfigDifferences(differences: Map<String, Any>) {
         configDifferences.clear()
         configDifferences.putAll(differences)
-        NdkPluginCaller.setStaticData(mapOf("config" to configDifferences))
+        NdkPluginCaller.setStaticData(mapOf("usage" to mapOf("config" to configDifferences)))
     }
 
     override fun setCallbackCounts(newCallbackCounts: Map<String, Int>) {
@@ -33,6 +46,16 @@ class InternalMetricsImpl : InternalMetrics {
     override fun notifyRemoveCallback(callback: String) {
         modifyCallback(callback, -1)
         NdkPluginCaller.notifyRemoveCallback(callback)
+    }
+
+    override fun setMetadataTrimMetrics(stringsTrimmed: Int, charsRemoved: Int) {
+        metadataStringsTrimmedCount = stringsTrimmed
+        metadataCharsTruncatedCount = charsRemoved
+    }
+
+    override fun setBreadcrumbTrimMetrics(breadcrumbsRemoved: Int, bytesRemoved: Int) {
+        breadcrumbsRemovedCount = breadcrumbsRemoved
+        breadcrumbBytesRemovedCount = bytesRemoved
     }
 
     private fun modifyCallback(callback: String, delta: Int) {