+++ bugsnag-plugin-react-native/src/test/java/com/bugsnag/android/TestData.java
@@ -40,6 +40,7 @@ class TestData {
                 32,
                 32,
                 1000,
+                10000,
                 LazyKt.lazy(new Function0<File>() {
                     @Override
                     public File invoke() {